requirements.txt
fastapi
PyMuPDF
uvicorn
pydantic
pandas
python-dotenv
python-multipart
httpx
scikit-learn
aiofiles
langchain
langchain-core
langchain-openai
torch
transformers
openai

